backend/requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]
python-multipart
aiofiles

# LLM & Embeddings - LangChain 1.0.5
langchain==1.0.5
langchain-core==1.0.4
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-openai==1.0.2
langchain-google-genai==3.0.2

# Vector Store & Embeddings
faiss-cpu
tiktoken

# LLM Providers
ollama==0.6.0
google-generativeai==0.7.0
openai>=1.0.0

# Data Processing
pymupdf
tqdm
pydantic>=2.0
python-dotenv
requests

# Optional but recommended
certifi